update: ciclo callback for logging to wandb
